Fix an extremely nasty cell generation bug
Fixed a very nasty bug that would cause generated cell polygons to
receive the wrong edges. The root of the issue was that the edges
within a span were sorted on the X axis by their bounding boxes in
order to be processed in the correct left-to-right order. This was
incorrect, however: The minimum and maximum X values of two edges
could be completely identical and they would therefore be processed
in an arbitrary order. Instead, the bounding boxes of the _segments
of the edges that intersect the span_ should have been used instead.